Refactor and restructure
